fixing duplication complaints
